+++ src/health/Ventricle.java
@@ -69,7 +69,7 @@ public class Ventricle
     /**
      * @return the volume
      */
-    public int getVolume()
+    public double getVolume()
     {
         return volume;
     }
